requirements.txt
setuptools
wheel
PyYAML
pytest
nmigen-yosys
edalize==0.2.3
fusesoc==1.11.0
git+https://github.com/nmigen/nmigen
